+++ src/collective/bcast.c
@@ -12,7 +12,7 @@ int bcast(void* buffer,
     MPI_Comm_rank(comm, &rank);
     MPI_Comm_size(comm, &num_procs);
 
-    int num_steps = log2(num_procs);
+    int num_steps = log2((float)(num_procs));
     int tag = 204857;
     MPI_Status status;
